qrc
Macro compress_data_macro

macro_rules! compress_data_macro {
    ($data:expr) => { ... };
}

Compresses data before encoding it into a QR code.

This is beneficial for encoding large amounts of data into a QR code by reducing the size of the data.

Parameters

$data:expr: The data to be compressed and encoded.

Example

use qrc::QRCode; // Import QRCode type
use qrc::compress_data_macro; // Corrected import to use the macro
let compressed_data = compress_data_macro!("Some large string of data"); // Correct usage