Macro qrc::create_dynamic_qr

macro_rules! create_dynamic_qr {
    ($initial_data:expr) => { ... };
}

Generates a dynamic QR code, which can be updated after creation.

Useful for scenarios where the data linked to the QR code might change over time, such as promotional offers or event details.

Parameters

$initial_data:expr: The initial data for the QR code.

Example

use qrc::QRCode; // Import QRCode type
use qrc::create_dynamic_qr;
create_dynamic_qr!("Initial Data");