Struct qrc::QRCode

#[non_exhaustive]
pub struct QRCode {
    pub data: Vec<u8>,
    /* private fields */
}

QRCode is a structure that contains data in the form of a vector of bytes.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

data: Vec<u8>

The data field holds the data to be encoded in the QR code.

Implementations

impl QRCode

Implementation of QRCode structure.

pub fn new(data: Vec<u8>) -> Self

Creates a new QRCode structure with the given data.

pub fn from_string(data: String) -> Self

The from_string method creates a new instance of the QRCode struct by converting the given string data into a vector of bytes

pub fn from_bytes(data: Vec<u8>) -> Self

Creates a new QRCode structure from a vector of bytes.

pub fn to_qrcode(&self) -> QrCode

Converts the QRCode structure to a QrCode structure.

pub fn to_png(&self, width: u32) -> ImageBuffer<Rgba<u8>, Vec<u8>>

Converts the QRCode structure to a PNG image.

pub fn to_jpg(&self, width: u32) -> ImageBuffer<Rgba<u8>, Vec<u8>>

Converts the QRCode structure to a JPG image.

pub fn to_gif(&self, width: u32) -> ImageBuffer<Rgba<u8>, Vec<u8>>

Converts the QRCode structure to a GIF image.

pub fn to_svg(&self, width: u32) -> String

Converts the QRCode structure to an SVG image.

pub fn colorize(&self, color: Rgba<u8>) -> RgbaImage

The colorize method creates a new PNG image of the QR code using the data stored in the QRCode and the given color value to colorize the QR code.

pub fn resize(&self, width: u32, height: u32) -> RgbaImage

The resize method creates a new PNG image of the QR code using the data stored in the QRCode and the given width and height values to resize the QR code.

pub fn add_image_watermark(img: &mut RgbaImage, watermark: &RgbaImage)

The add_image_watermark method adds a watermark to the given image.

pub fn create_multilanguage(data_map: HashMap<String, String>) -> Self

Creates a QR code with multi-language support.

pub fn create_dynamic(initial_data: &str) -> Self

Generates a dynamic QR code, which can be updated after creation.

pub fn combine_qr_codes(codes: Vec<QRCode>) -> Result<Self, &'static str>

Combines multiple QR codes into a single larger QR code.

pub fn compress_data(data: &str) -> Vec<u8>

Compresses data before encoding it into a QR code.

pub fn batch_generate_qr_codes(data: Vec<String>) -> Vec<QRCode>

Decompresses data after decoding it from a QR code.

pub fn overlay_image(&self, overlay: &RgbaImage) -> RgbaImage

Implementation for overlaying an image on top of a QR code.

pub fn set_encoding_format(&self, format: &str) -> Result<Self, &'static str>

Sets the encoding format for the QR code and returns a new instance.

pub fn get_encoding_format(&self) -> &str

Gets the encoding format of the QR code.