Macro qrc::combine_qr_codes

macro_rules! combine_qr_codes {
    ($codes:expr) => { ... };
}

Combines multiple QR codes into a single QR code.

This macro is useful for scenarios where multiple QR codes need to be combined, such as creating a composite QR code with several data sources.

Parameters

An array of QRCode instances to combine.

Example

use qrc::QRCode; // Import QRCode type
use qrc::combine_qr_codes;

let qr_code1 = QRCode::from_string("Data 1".to_string());
let qr_code2 = QRCode::from_string("Data 2".to_string());
let qr_code3 = QRCode::from_string("Data 3".to_string());

let combined_qr_code = combine_qr_codes!(vec![qr_code1, qr_code2, qr_code3]);