qrc
Macro resize

macro_rules! resize {
    ($qrcode:expr, $size:expr) => { ... };
}

Sets the size of the QR code.

This macro allows the user to specify the size of the QR code. The size is typically defined in terms of pixels or modules (the small squares that make up a QR code).

Parameters

$qrcode:expr: An instance of QRCode.

$size:expr: The desired size for the QR code.

Example

use qrc::QRCode;
use qrc::resize;

let qrcode = QRCode::new("Hello, world!".as_bytes().to_vec());
let resized_qrcode = resize!(qrcode, 256);