Struct qrc::QRCode

#[non_exhaustive]
pub struct QRCode {
    pub data: Vec<u8>,
    /* private fields */
}

Represents a QR code containing data.

This struct can be used to generate QR code images in various formats. It supports encoding data as a QR code and rendering it in formats like PNG, JPG, and SVG.

Examples

use qrc::QRCode;

// Create a new QR code with text data
let qr = QRCode::new("Hello, world!".as_bytes().to_vec());

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

data: Vec<u8>

The data field holds the data to be encoded in the QR code.

Implementations

impl QRCode

Implementation of QRCode structure.

pub fn new(data: Vec<u8>) -> Self

Creates a new QRCode instance with the given data.

Examples

use qrc::QRCode;

let qr = QRCode::new("Hello, world!".as_bytes().to_vec());

Parameters

data: A Vec<u8> representing the data to be encoded in the QR code.

Returns

A new QRCode instance.

pub fn from_string(data: String) -> Self

The from_string method creates a new instance of the QRCode struct by converting the given string data into a vector of bytes

pub fn from_bytes(data: Vec<u8>) -> Self

Creates a new QRCode structure from a vector of bytes.

pub fn to_qrcode(&self) -> QrCode

Converts the QRCode structure to a QrCode structure.

pub fn to_png(&self, width: u32) -> ImageBuffer<Rgba<u8>, Vec<u8>>

Converts the QRCode structure to a PNG image.

Examples

use qrc::QRCode;

// Convert a string slice to a String using `.to_string()`
let qr = QRCode::from_string("Hello, world!".to_string());
let png_image = qr.to_png(256);

Parameters

width: The width of the image in pixels.

Returns

An ImageBuffer representing the QR code in PNG format.

pub fn to_jpg(&self, width: u32) -> ImageBuffer<Rgba<u8>, Vec<u8>>

Converts the QRCode structure to a JPG image.

Parameters

width: The width of the desired image in pixels.

Returns

An ImageBuffer representing the QR code in JPG format.

pub fn to_gif(&self, width: u32) -> ImageBuffer<Rgba<u8>, Vec<u8>>

Converts the QRCode structure to a GIF image.

Parameters

width: The width of the desired image in pixels.

Returns

An ImageBuffer representing the QR code in GIF format.

pub fn to_svg(&self, width: u32) -> String

Converts the QRCode structure to an SVG image.

Parameters

width: The width of the desired image in pixels.

Returns

A String representing the QR code in SVG format.

pub fn colorize(&self, color: Rgba<u8>) -> RgbaImage

Colorizes the QR code with the specified color.

Parameters

color: The Rgba<u8> color value to use for the QR code.

Returns

A colorized RgbaImage of the QR code.

pub fn resize(&self, width: u32, height: u32) -> RgbaImage

Resizes the QR code image to the specified width and height.

Parameters

width: The width of the image in pixels.

height: The height of the image in pixels.

Returns

A resized RgbaImage of the QR code.

pub fn add_image_watermark(img: &mut RgbaImage, watermark: &RgbaImage)

Adds a watermark image to the QR code.

Parameters

img: A mutable reference to the RgbaImage of the QR code.

watermark: A reference to the watermark RgbaImage.

pub fn create_multilanguage(data_map: HashMap<String, String>) -> Self

Creates a multilingual QR code based on a map of language codes to data strings.

Parameters

data_map: A HashMap mapping language codes (String) to data (String).

Returns

A QRCode instance representing a multilingual QR code.

pub fn create_dynamic(initial_data: &str) -> Self

Generates a dynamic QR code that can be updated after creation.

Parameters

initial_data: A string slice representing the initial data for the QR code.

Returns

A QRCode instance representing a dynamic QR code.

pub fn combine_qr_codes(codes: Vec<QRCode>) -> Result<Self, &'static str>

Combines multiple QR codes into a single larger QR code.

Parameters

codes: A vector of QRCode instances to combine.

Returns

A Result which is either a combined QRCode instance or an error string.

pub fn compress_data(data: &str) -> Vec<u8>

Compresses the provided data string using Zlib compression.

Parameters

data: A string slice representing the data to compress.

Returns

A Vec<u8> containing the compressed data.

pub fn batch_generate_qr_codes(data: Vec<String>) -> Vec<QRCode>

Generates a batch of QR codes from a vector of data strings.

Parameters

data: A vector of strings, each representing data for a separate QR code.

Returns

A vector of QRCode instances.

pub fn overlay_image(&self, overlay: &RgbaImage) -> RgbaImage

Overlays an image on top of the QR code.

Parameters

overlay: A reference to the RgbaImage to overlay on the QR code.

Returns

A combined RgbaImage with the overlay applied.

pub fn set_encoding_format(&self, format: &str) -> Result<Self, &'static str>

Sets the encoding format of the QR code.

Parameters

format: A string slice representing the encoding format.

Returns

A Result which is either a new QRCode instance with updated encoding or an error string.

pub fn get_encoding_format(&self) -> &str

Retrieves the encoding format of the QR code.

Returns

A string slice representing the encoding format.